Fix incorrect header dependencies
Installed headers cannot depend on noinst headers.
