+++ Package.swift
@@ -524,11 +524,11 @@ let headers: [String] = [
     "realm/sync/changeset_encoder.hpp",
     "realm/sync/changeset_parser.hpp",
     "realm/sync/client.hpp",
+    "realm/sync/client_base.hpp",
     "realm/sync/config.hpp",
     "realm/sync/history.hpp",
     "realm/sync/impl/clamped_hex_dump.hpp",
     "realm/sync/impl/clock.hpp",
-    "realm/sync/inspector/util.hpp",
     "realm/sync/instruction_applier.hpp",
     "realm/sync/instruction_replication.hpp",
     "realm/sync/instructions.hpp",
